refactor: updating helpers function
